normal pipeline fixes
This fixes

gets the latest AA-backend image tag. Previously this was using qa, but the image pull policy is "pull if not present", so the image was in some cases 2 months old on the server.
"set +x" prevents the script from failing on command error
dev-prod renamed keycloack route to keycloack-mocks. this is also fixed