Add removeDefault to update function of QueryParams
Allows to remove all params from default namespace in URL params. This is needed by Reports page because if the selected report changes we need to remove all its settings saved in the URL.
Given we are limiting number of API calls we have to this change at once otherwise only one will be done (either selected report or removing of params). That's why it's not possible to do it by chnaging URL directly or calling update function twice.